Move around some state code to a better location
